requirements.txt
CPU-G==0.11.1
Django==1.8.18
Flask==0.11.1
JSON-log-formatter==0.1.0
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
PyJWT==1.6.4
PyQRCode==1.2.1
PyYAML==3.10
Pygments==2.2.0
SecretStorage==2.0.0
SwampDragon==0.4.2.2
Twisted-Core==13.2.0
Twisted-Web==13.2.0
VTK==5.8.0
Werkzeug==0.11.11
XlsxWriter==0.9.3
adium-theme-ubuntu==0.3.4
apptools==4.4.0
apt-xapian-index==0.45
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
boto==2.20.1
bzr==2.7.0.dev1
certifi==2016.9.26
cffi==1.8.3
chardet==2.0.1
click==6.6
colorama==0.2.5
command-not-found==0.3
configobj==4.7.2
cryptography==1.5.2
cssselect==1.0.1
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
django-admin-bootstrapped==2.5.7
django-annoying==0.8.1
django-appconf==1.0.2
django-bootstrap-toolkit==2.15.0
django-bootstrap3==5.4.0
django-bootstrap3-datepicker==0.3.1
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.4.0
django-datatable==0.1.5
django-filetransfers==0.1.0
django-render-block==0.5
django-select2==6.1.0
django-simple-captcha==0.4.4
django-tables2==1.0.4
django-templated-email==2.2.0
django-widget-tweaks==1.4.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
dnspython==1.11.1
dulwich==0.9.4
duplicity==0.6.23
enum34==1.1.6
fastimport==0.9.2
functools32==3.2.3.post2
gTTS==1.2.0
gTTS-token==1.1.1
grammar-check==1.3.1
gyp==0.1
html5lib==0.999
httplib2==0.8
idna==2.1
imageio==2.2.0
ipaddress==1.0.17
ipython==1.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jsonschema==2.6.0
jupyter-core==4.3.0
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
lxml==3.3.3
matplotlib==1.3.1
mayavi==4.5.0
mlab==1.1.4
nbformat==4.3.0
nltk==3.2.4
numpy==1.12.1
oauth==1.0.1
oauthlib==0.6.1
oneconf==0.3.7.14.4.1
paramiko==1.10.1
pexpect==3.1
piston-mini-client==0.7.5
plotly==2.0.8
pyOpenSSL==0.13
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pycups==1.9.66
pyface==5.1.0
pygobject==3.12.0
pyinotify==0.9.4
pymssql==2.1.3
pyparsing==2.0.1
pypng==0.0.18
pyserial==2.6
pysmbc==1.0.14.1
python-apt===0.9.3.5ubuntu2
python-cloudfiles==1.7.10
python-dateutil==2.5.3
python-debian===0.1.21-nmu2ubuntu2
python-http-client==3.0.0
pytz==2017.2
pyxdg==0.25
rabbitvcs==0.16.0
redis==2.10.5
reportlab==3.0
requests==2.2.1
sendgrid==3.6.5
sendgrid-django==4.1.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.3.1
singledispatch==3.4.0.3
six==1.10.0
sockjs-tornado==1.0.3
software-center-aptd-plugins==0.0.0
ssh-import-id==3.21
suds==0.4.1
system-service==0.1.6
tornado==4.4.2
tornado-redis==2.4.18
traitlets==4.3.2
traits==4.6.0
traitsui==5.1.0
unity-lens-photos==1.0
urllib3==1.7.1
wadllib==1.3.2
wheel==0.24.0
wsgiref==0.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose===3.6.3build2
xlrd==1.0.0
zope.interface==4.0.5
